Moves SourceLauncherService.java into judge/services
